requirements.txt
python-dotenv
langchain
langchain-community
langchain-core
langchain-ollama
langchain-groq
faiss-cpu
pdfplumber
